Move common OTLP to a separate file
