+++ pkg/otlp/otlp.go
@@ -18,12 +18,9 @@
 package otlp
 
 import (
-	"bytes"
-	"compress/gzip"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io"
 
 	"github.com/siglens/siglens/pkg/config"
 	"github.com/siglens/siglens/pkg/es/writer"
@@ -37,20 +34,8 @@ import (
 	coltracepb "go.opentelemetry.io/proto/otlp/collector/trace/v1"
 	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 	tracepb "go.opentelemetry.io/proto/otlp/trace/v1"
-	"google.golang.org/genproto/googleapis/rpc/status"
 	"google.golang.org/protobuf/proto"
 )
-
-const contentTypeHeader = "Content-Type"
-const protobufContentType = "application/x-protobuf" // This might be OTLP specific.
-
-func getContentType(ctx *fasthttp.RequestCtx) string {
-	return string(ctx.Request.Header.Peek(contentTypeHeader))
-}
-
-func setContentType(ctx *fasthttp.RequestCtx, contentType string) {
-	ctx.Response.Header.Set(contentTypeHeader, contentType)
-}
 
 func ProcessTraceIngest(ctx *fasthttp.RequestCtx, myid int64) {
 	if hook := hooks.GlobalHooks.OverrideIngestRequestHook; hook != nil {
@@ -144,36 +129,6 @@ func ProcessTraceIngest(ctx *fasthttp.RequestCtx, myid int64) {
 	usageStats.UpdateTracesStats(uint64(len(data)), uint64(numSpans), myid)
 	// Send the appropriate response.
 	handleTraceIngestionResponse(ctx, numSpans, numFailedSpans)
-}
-
-func getUncompressedData(ctx *fasthttp.RequestCtx) ([]byte, error) {
-	data := ctx.PostBody()
-	if requiresGzipDecompression(ctx) {
-		reader, err := gzip.NewReader(bytes.NewReader(data))
-		if err != nil {
-			return nil, fmt.Errorf("cannot gzip decompress: err=%v", err)
-		}
-
-		data, err = io.ReadAll(reader)
-		if err != nil {
-			return nil, fmt.Errorf("cannot gzip decompress: err=%v", err)
-		}
-	}
-
-	return data, nil
-}
-
-func requiresGzipDecompression(ctx *fasthttp.RequestCtx) bool {
-	encoding := string(ctx.Request.Header.Peek("Content-Encoding"))
-	if encoding == "gzip" {
-		return true
-	}
-
-	if encoding != "" && encoding != "none" {
-		log.Errorf("requiresGzipDecompression: invalid content encoding: %s. Request headers: %v", encoding, ctx.Request.Header.String())
-	}
-
-	return false
 }
 
 func unmarshalTraceRequest(data []byte) (*coltracepb.ExportTraceServiceRequest, error) {
@@ -309,24 +264,6 @@ func linksToJson(spanLinks []*tracepb.Span_Link) ([]byte, error) {
 	}
 
 	return jsonLinks, nil
-}
-
-func setFailureResponse(ctx *fasthttp.RequestCtx, statusCode int, message string) {
-	ctx.SetStatusCode(statusCode)
-
-	failureStatus := status.Status{
-		Code:    int32(statusCode),
-		Message: message,
-	}
-
-	bytes, err := proto.Marshal(&failureStatus)
-	if err != nil {
-		log.Errorf("setFailureResponse: failed to marshal failure status. err: %v. Status: %+v", err, &failureStatus)
-	}
-	_, err = ctx.Write(bytes)
-	if err != nil {
-		log.Errorf("sendFailureResponse: failed to write failure status: %v", err)
-	}
 }
 
 func handleTraceIngestionResponse(ctx *fasthttp.RequestCtx, numSpans int, numFailedSpans int) {
